Added support and rewrite url rules for english URLs

+++ src/js/ssbHeader.js
@@ -29,9 +29,31 @@ class SSBHeader extends Component {
         e.preventDefault();
         const lang = e.currentTarget.lang;
 
-        sessionStorage.clear('selectedAPILanguage')
-        sessionStorage.setItem('selectedLanguage', lang)
-        window.location.reload()
+
+
+
+        // regex to make sure current url does not contain "en/" in its url and groups for easy insert of "en/"
+        // ex(http(s)://www.ssb.no/)(klass/)(#/)(klassifikasjoner/7)
+        const NORWEGIAN_URL_REGEX = /(http(s?):\/\/.*?\/)(?!en\/)([^#\/]*\/)?(#\/)?(.*)/;
+
+        // regex to make sure current url is an english one (contains "en/") and groups for easy removal of "en/"
+        // ex(http(s)://www.ssb.no/)(en/)(#/)(klassifikasjoner/7)
+        let ENGLISH_URL_REGEX = /(http(s?):\/\/.*?)(en\/)(#\/)?(.*)/;
+
+        if (document.URL.match(ENGLISH_URL_REGEX) ) {
+            //remove (en/) group and navigate to new url
+            let location = document.URL.replace(ENGLISH_URL_REGEX, "$1$4$5");
+            window.location = location
+        }else if (document.URL.match(NORWEGIAN_URL_REGEX) ){
+            // insert "en/" in url
+            let location = document.URL.replace(NORWEGIAN_URL_REGEX, "$1en/$3#/$5");
+            window.location = location
+        } else {
+            console.error("no regex url match")
+            sessionStorage.clear('selectedAPILanguage')
+            sessionStorage.setItem('selectedLanguage', lang)
+             window.location.reload()
+        }
     }
 
     renderMainMenu() {
